--- a/data_key_test.go
+++ b/data_key_test.go
@@ -3,6 +3,8 @@ package mmdbwriter
 import (
 	"testing"
 
+	"github.com/stretchr/testify/assert"
+
 	"github.com/maxmind/mmdbwriter/mmdbtype"
 )
 
@@ -36,4 +38,30 @@ func BenchmarkKeyGeneration(b *testing.B) {
 			}
 		}
 	})
+}
+
+func TestKeyWriter(t *testing.T) {
+	assertions := assert.New(t)
+	writer := newKeyWriter()
+
+	valueOne := mmdbtype.String("some test value to be turned into a key")
+	valueTwo := mmdbtype.String("another test value to be turned into a key")
+
+	keyOne, err := writer.key(valueOne)
+
+	assertions.NoErrorf(err, "expected no error")
+
+	keyTwo, err := writer.key(valueTwo)
+
+	assertions.NoErrorf(err, "expected no error")
+
+	// The keys should be uniformly distributed, so the change of this test breaking at random is 1 in 2^64.
+	assertions.NotEqual(keyOne, keyTwo, "expected keys to be different")
+
+	keyOneAgain, err := writer.key(valueOne)
+
+	assertions.NoErrorf(err, "expected no error")
+
+	// Test this after keyTwo was created to make sure the state is not being reused.
+	assertions.Equal(keyOne, keyOneAgain, "expected keys to be the same for the same value")
 }