--- a/data_map_test.go
+++ b/data_map_test.go
@@ -3,9 +3,10 @@ package mmdbwriter
 import (
 	"testing"
 
-	"github.com/maxmind/mmdbwriter/mmdbtype"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+
+	"github.com/maxmind/mmdbwriter/mmdbtype"
 )
 
 func TestDataMap(t *testing.T) {
@@ -19,9 +20,8 @@ func TestDataMap(t *testing.T) {
 	assert.Equal(
 		t,
 		&dataMapValue{
-			data: v,
-			key: "\x87\x02\xf53\x8b\x96\xfdǻQ\x97\x9c\xe2\xcc\\\xda\xf2\xb1\xd7" +
-				"\xc1L\xc5l\xfd\x83\xfc\x97\xd6\x03\xf5\xedr",
+			data:     v,
+			key:      123456789,
 			refCount: 1,
 		},
 		dmv,