Two changes:
- Changed the representation of the internal `dataMapKey` to a `uint64` instead of a `string`.
- Changed the `keyWriter` to use a [`maphash.Hash`](https://pkg.go.dev/hash/maphash#Hash) instead of a sha256 since there is no need for cryptographically secure hashing on a deduplication key. This could be made configurable if there is actually such a use case.

As noted in the comment, I checked the effect of this change on the size and alignment of `dataMapValue`. The alignment remains at 8 bytes, and the size went down from 40 to 32 bytes due to `string` requiring two words. A test was added to log these properties since those should be kept under control.

Also includes a benchmark for the new `keyWriter.key` method. Here are the results from the original implementation, the one proposed by this PR and a few others that I tried:
```
# Result using buffered sha256
BenchmarkKeyGeneration/simple_value-12            3940114               299.3 ns/op            64 B/op          3 allocs/op
BenchmarkKeyGeneration/nested_value-12            1000000              1052 ns/op             272 B/op         14 allocs/op

# Result using unbuffered fnv.New64 and bytes return type
BenchmarkKeyGeneration/simple_value-12            6729507               188.6 ns/op            96 B/op          6 allocs/op
BenchmarkKeyGeneration/nested_value-12            1015862              1166 ns/op             360 B/op         39 allocs/op

# Result using buffered fnv.New64 and uint64 return type
BenchmarkKeyGeneration/simple_value-12           10141742               114.7 ns/op            32 B/op          2 allocs/op
BenchmarkKeyGeneration/nested_value-12            1547077               777.9 ns/op           240 B/op         13 allocs/op

# Result using maphash.Hash (internally buffered) and uint64 return type
BenchmarkKeyGeneration/simple_value-12           14879239                78.18 ns/op           32 B/op          2 allocs/op
BenchmarkKeyGeneration/nested_value-12            1629522               714.8 ns/op           240 B/op         13 allocs/op
```

The change causes a 30% improvement for nested values and up to 70% improvement for simple values.

Also a point of notice, converting a `[]byte` to `string` causes a new allocation and copy of the bytes every time ([example on playground](https://go.dev/play/p/QkwpY5Bmixk)) while using a `uint64` makes the conversion to `dataMapKey` a noop (see [here](https://godbolt.org/z/edrszvfqx) that it is gone after compilation). So this change also makes the other functions that use the key more efficient.