@oschwald 2022-07-25
My primary concern is the risk of collisions. In terms of benefit, I think there would be a much more significant benefit if we wrote our own hash function so that we did not need to convert the data to `[]byte` at all. This would need to be done carefully though.

Also, I believe the conversion from `[]byte` to `string` is minimal as Go has an optimization to avoid the conversion for lookups like `datamap[string(byteSlice)]`. It would only be the inserts causing allocations.