@Fryuni 2022-07-25
You are right about that, I didn't consider the number of expected collisions.
Using the Birthday paradox formula for 8 million nodes (the size of GeoLite2-City), a 64 bits address space, and about 20 values per node (nested maps), we should expect to have about 100k collisions. Which is a huge problem.

> Also, I believe the conversion from []byte to string is minimal as Go has an optimization to avoid the conversion for lookups like datamap[string(byteSlice)]. It would only be the inserts causing allocations.

The problem with using strings is that the map will store pointers, so every key will have to go through an extra indirection when reading _and_ an extra allocation when inserting.

This function is called for every data point on every node. In our case, that is hundreds of millions of times, so this change alone saves more than 2 minutes of runtime per run.
We could use a 128 bits value for the hash, in which case it is only expected to have a collision with more than 10^11 nodes.
A 128 bits hash (16 bytes) would still be smaller than a slice of bytes or a string (both with a header of 24 bytes) and would not require an extra indirection.

I'll do some more in-depth profiling to see how big of an impact using a 128 bits hash instead of a byte slice would have on both runtime and memory use. Especially memory spikes caused by the multiple small allocations.

Thanks for pointing the collision problem.