@Fryuni 2022-08-10
Hey @oschwald. I profiled our application for memory allocations/use and CPU time, and more than half of it was spent on the insert operation for the writer.
We have to hit 5 databases and quite an expensive computation to get each value we are storing, and the insert operation itself is still taking as much time as that.

Do you mind if I open a tracking issue to log what I find and what optimizations could be done here? I'm probably gonna invest some time on this over the next few weeks.

I can close this PR in the meantime, I found that this improvement is only significant on my synthetic benchmark where only a single value is serialized multiple times.