@oschwald 2022-08-10
@Fryuni, that sounds good. Thanks! I do have some thoughts on performance improvements based on our usage, but our usage patterns sound a bit different. It would be good to have a general issue to discuss this more.